Consolidate enums to prevent circular dependencies
