+++ app/controllers/reply.js
@@ -1,16 +1,17 @@
 import wizard from '@x-govuk/govuk-prototype-wizard'
 
-import { GillickCompetent } from '../models/gillick.js'
+import {
+  GillickCompetent,
+  ReplyDecision,
+  ReplyMethod,
+  ReplyRefusal,
+  VaccinationOutcome
+} from '../enums.js'
 import { Parent } from '../models/parent.js'
 import { PatientSession } from '../models/patient-session.js'
 import { Programme } from '../models/programme.js'
-import {
-  Reply,
-  ReplyDecision,
-  ReplyMethod,
-  ReplyRefusal
-} from '../models/reply.js'
-import { Vaccination, VaccinationOutcome } from '../models/vaccination.js'
+import { Reply } from '../models/reply.js'
+import { Vaccination } from '../models/vaccination.js'
 import { today } from '../utils/date.js'
 import { hasAnswersNeedingTriage } from '../utils/reply.js'
 import { formatParent } from '../utils/string.js'
