+++ app/generators/consent.js
@@ -1,9 +1,14 @@
 import { fakerEN_GB as faker } from '@faker-js/faker'
 
 import { healthConditions } from '../datasets/health-conditions.js'
+import {
+  ProgrammeType,
+  ReplyDecision,
+  ReplyMethod,
+  ReplyRefusal
+} from '../enums.js'
 import { Consent } from '../models/consent.js'
-import { ProgrammeType } from '../models/programme.js'
-import { ReplyDecision, ReplyMethod, ReplyRefusal } from '../models/reply.js'
+import {} from '../models/programme.js'
 import { removeDays, today } from '../utils/date.js'
 import {
   getHealthAnswers,
