+++ app/generators/user.js
@@ -1,6 +1,7 @@
 import { fakerEN_GB as faker } from '@faker-js/faker'
 
-import { User, UserRole } from '../models/user.js'
+import { UserRole } from '../enums.js'
+import { User } from '../models/user.js'
 
 /**
  * Generate fake user
