+++ app/utils/reply.js
@@ -2,11 +2,14 @@ import { faker } from '@faker-js/faker'
 import _ from 'lodash'
 
 import { healthConditions } from '../datasets/health-conditions.js'
+import {
+  ConsentOutcome,
+  ParentalRelationship,
+  ProgrammeType,
+  ReplyDecision,
+  ReplyRefusal
+} from '../enums.js'
 import { Child } from '../models/child.js'
-import { ParentalRelationship } from '../models/parent.js'
-import { ConsentOutcome } from '../models/patient-session.js'
-import { ProgrammeType } from '../models/programme.js'
-import { ReplyDecision, ReplyRefusal } from '../models/reply.js'
 
 import { formatParentalRelationship } from './string.js'
