+++ lib/create-data.js
@@ -9,6 +9,22 @@ import organisationsData from '../app/datasets/organisations.js'
 import schoolsData from '../app/datasets/schools.js'
 import usersData from '../app/datasets/users.js'
 import vaccinesData from '../app/datasets/vaccines.js'
+import {
+  AcademicYear,
+  ConsentOutcome,
+  ConsentWindow,
+  PatientOutcome,
+  ProgrammePreset,
+  ProgrammeType,
+  NoticeType,
+  MoveSource,
+  RegistrationOutcome,
+  SchoolPhase,
+  ScreenOutcome,
+  SessionType,
+  UploadType,
+  UserRole
+} from '../app/enums.js'
 import { generateBatch } from '../app/generators/batch.js'
 import { generateCohort } from '../app/generators/cohort.js'
 import { generateConsent } from '../app/generators/consent.js'
@@ -23,31 +39,14 @@ import { generateUpload } from '../app/generators/upload.js'
 import { generateUser } from '../app/generators/user.js'
 import { generateVaccination } from '../app/generators/vaccination.js'
 import { Clinic } from '../app/models/clinic.js'
-import { AcademicYear, Cohort } from '../app/models/cohort.js'
+import { Cohort } from '../app/models/cohort.js'
 import { Instruction } from '../app/models/instruction.js'
-import { Move, MoveSource } from '../app/models/move.js'
-import { NoticeType } from '../app/models/notice.js'
+import { Move } from '../app/models/move.js'
 import { Organisation } from '../app/models/organisation.js'
-import {
-  ConsentOutcome,
-  PatientOutcome,
-  PatientSession,
-  ScreenOutcome
-} from '../app/models/patient-session.js'
-import {
-  ProgrammePreset,
-  ProgrammeType,
-  programmeTypes
-} from '../app/models/programme.js'
-import { SchoolPhase } from '../app/models/school.js'
-import {
-  ConsentWindow,
-  RegistrationOutcome,
-  Session,
-  SessionType
-} from '../app/models/session.js'
-import { UploadType } from '../app/models/upload.js'
-import { User, UserRole } from '../app/models/user.js'
+import { PatientSession } from '../app/models/patient-session.js'
+import { programmeTypes } from '../app/models/programme.js'
+import { Session } from '../app/models/session.js'
+import { User } from '../app/models/user.js'
 import { Vaccination } from '../app/models/vaccination.js'
 import {
   addDays,
